Uploading new example!
Example for the UART category, useful for testing UART0 or UART1 in the RP2040. It is helpful for internal testing by connecting TXD (transmit data) to RxD (receive data), for example, for UART0, GPIO(0) is connected to GPIO(1), and then the message is sent to be transmitted and read on the same port.